MHV / Refactor UHD service (#24223)
* extract labs and tests methods into adapter, adjust serializer + usage, update all tests

* set up a UHD client; make labs use client

* conditions use new client

* notes use new client

* prescriptions use client

* fix Rx tests (mostly); change single note method back

* remove old serializer from codeowners file

* single condition method uses client

* add type back into lab stuff for logging

* no v2 on notes adapter

* update type in serializer

* remove byebug


* adjust service spec to use client for LOINC code logging tests

* add V2 back into notes adapter for now

* tidying

* remove V2 from notes adapter for consistency

* tidying, add info for running betamocks, fix test after merge from master

---------
