+++ lib/unified_health_data/README.md
@@ -8,19 +8,27 @@ The Unified Health Data service provides a unified interface to retrieve health 
 lib/unified_health_data/
 ├── README.md                    # This file
 ├── service.rb                   # Main service class
+├── client.rb                   # Main client class
 ├── configuration.rb             # Service configuration
 ├── reference_range_formatter.rb # Lab result formatting utilities
 ├── models/                      # Data models using Vets::Model
 │   ├── clinical_notes.rb       # Clinical notes model
+│   ├── condition.rb             # Condition model
 │   ├── lab_or_test.rb          # Lab results and tests model
 │   └── prescription.rb         # Prescription model
 ├── adapters/                    # Data source adapters
-│   ├── clinical_notes_adapter.rb          # Clinical notes parser (V2 namespace)
+│   ├── clinical_notes_adapter.rb          # Clinical notes parser 
+│   ├── conditions_adapter.rb          # Conditions parser 
+│   ├── labs_or_test_adapter.rb          # Labs and tests parser
 │   ├── prescriptions_adapter.rb           # Main prescription parser
 │   ├── vista_prescription_adapter.rb      # VistA medication parser
 │   └── oracle_health_prescription_adapter.rb # Oracle Health FHIR parser
 ├── serializers/                # API response serializers
-│   └── clinical_notes_serializer.rb       # Clinical notes serialization
+│   ├── clinical_notes_serializer.rb       # Clinical notes serialization
+│   ├── conditions_serializer.rb          # Conditions serialization 
+│   ├── labs_or_test_serializer.rb          # Labs and tests serialization
+│   ├── prescriptions_serializer.rb           # Main prescription serialization
+│   └── prescriptions_refills_serializer.rb     # Prescription refills serialization
 └── logging.rb                  # Logging utilities
 ```
 
@@ -273,8 +281,52 @@ result = vista_adapter.parse(vista_medication_data)
 oracle_adapter = UnifiedHealthData::Adapters::OracleHealthPrescriptionAdapter.new
 result = oracle_adapter.parse(fhir_medication_request)
 
-# Test clinical notes adapter (V2 namespace)
-notes_adapter = UnifiedHealthData::V2::Adapters::ClinicalNotesAdapter.new
+# Test clinical notes adapter 
+notes_adapter = UnifiedHealthData::Adapters::ClinicalNotesAdapter.new
 result = notes_adapter.parse(document_reference_data)
 ```
 
+## Running the API locally
+
+In order to run the API locally, you can bypass authentication by adding a stub `test_user` and skipping authentication in the controller you are testing.
+
+For example, if you want to test the `labs_and_tests_controller` in the `my_health` module (`modules/my_health/app/controllers/my_health/v2/labs_and_tests_controller.rb`), it would look like this:
+
+```ruby
+module MyHealth
+  module V2
+    class LabsAndTestsController < ApplicationController
+      service_tag 'mhv-medical-records'
+
+      skip_before_action :authenticate # Add this
+
+      def index
+        start_date = params[:start_date]
+        end_date = params[:end_date]
+        labs = service.get_labs(start_date:, end_date:)
+        serialized_labs = UnifiedHealthData::LabOrTestSerializer.new(labs).serializable_hash[:data]
+        render json: serialized_labs,
+               status: :ok
+      end
+
+      private
+
+      def service
+      # Add a test user
+        test_user = OpenStruct.new( 
+          mhv_correlation_id: REDACTED, # Get the MHV ID of a test user
+          icn: 'REDACTED', # Get the ICN of a test user
+          edipi: '10000000',
+          last_name: 'Tester',
+          uuid: '12345',
+          flipper_id: '12345'
+        )
+        # Instantiate the service with your `test_user` rather than the `@current_user`
+        UnifiedHealthData::Service.new(test_user)
+      end
+    end
+  end
+end
+```
+
+You can then hit the expected endpoint (either in the browser or from a locally running vets-website). If using betamocks in the `vets-api-mockdata` repo, ensure you have a `default.yml` response added.
